+++ x-pack/plugins/lists/common/schemas/types/entries.ts
@@ -8,59 +8,11 @@
 
 import * as t from 'io-ts';
 
-import { NonEmptyString } from '../../siem_common_deps';
-import { operator, type } from '../common/schemas';
-
-import { nonEmptyOrNullableStringArray } from './non_empty_or_nullable_string_array';
-import { nonEmptyNestedEntriesArray } from './non_empty_nested_entries_array';
-
-export const entriesMatch = t.exact(
-  t.type({
-    field: NonEmptyString,
-    operator,
-    type: t.keyof({ match: null }),
-    value: NonEmptyString,
-  })
-);
-export type EntryMatch = t.TypeOf<typeof entriesMatch>;
-
-export const entriesMatchAny = t.exact(
-  t.type({
-    field: NonEmptyString,
-    operator,
-    type: t.keyof({ match_any: null }),
-    value: nonEmptyOrNullableStringArray,
-  })
-);
-export type EntryMatchAny = t.TypeOf<typeof entriesMatchAny>;
-
-export const entriesList = t.exact(
-  t.type({
-    field: NonEmptyString,
-    list: t.exact(t.type({ id: NonEmptyString, type })),
-    operator,
-    type: t.keyof({ list: null }),
-  })
-);
-export type EntryList = t.TypeOf<typeof entriesList>;
-
-export const entriesExists = t.exact(
-  t.type({
-    field: NonEmptyString,
-    operator,
-    type: t.keyof({ exists: null }),
-  })
-);
-export type EntryExists = t.TypeOf<typeof entriesExists>;
-
-export const entriesNested = t.exact(
-  t.type({
-    entries: nonEmptyNestedEntriesArray,
-    field: NonEmptyString,
-    type: t.keyof({ nested: null }),
-  })
-);
-export type EntryNested = t.TypeOf<typeof entriesNested>;
+import { entriesMatchAny } from './entry_match_any';
+import { entriesMatch } from './entry_match';
+import { entriesExists } from './entry_exists';
+import { entriesList } from './entry_list';
+import { entriesNested } from './entry_nested';
 
 export const entry = t.union([entriesMatch, entriesMatchAny, entriesList, entriesExists]);
 export type Entry = t.TypeOf<typeof entry>;
@@ -69,9 +21,6 @@ export const entriesArray = t.array(
   t.union([entriesMatch, entriesMatchAny, entriesList, entriesExists, entriesNested])
 );
 export type EntriesArray = t.TypeOf<typeof entriesArray>;
-
-export const nestedEntriesArray = t.array(t.union([entriesMatch, entriesMatchAny, entriesExists]));
-export type NestedEntriesArray = t.TypeOf<typeof nestedEntriesArray>;
 
 export const entriesArrayOrUndefined = t.union([entriesArray, t.undefined]);
 export type EntriesArrayOrUndefined = t.TypeOf<typeof entriesArrayOrUndefined>;
