+++ x-pack/plugins/lists/common/schemas/types/index.ts
@@ -11,6 +11,11 @@ export * from './default_create_comments_array';
 export * from './default_update_comments_array';
 export * from './default_namespace';
 export * from './entries';
+export * from './entry_match';
+export * from './entry_match_any';
+export * from './entry_list';
+export * from './entry_exists';
+export * from './entry_nested';
 export * from './non_empty_entries_array';
 export * from './non_empty_or_nullable_string_array';
 export * from './non_empty_nested_entries_array';
